pyproject.toml
[project]
name = "movie-finder-chain"
version = "0.1.0"
description = "LangGraph pipeline for movie discovery and Q&A"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "imdbapi-client",
    "langgraph>=0.2.0",
    "langchain-anthropic>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-core>=0.3.0",
    "langsmith>=0.1.0",
    "qdrant-client>=1.17.1",
    "openai>=2.29.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
lint = [
    "ruff>=0.9.0",
    "mypy>=1.14.0",
]
test = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-mock>=3.14.0",
    "pytest-cov>=6.0.0",
]
dev = [
    { include-group = "lint" },
    { include-group = "test" },
]

[tool.uv]
default-groups = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
imdbapi-client = { workspace = true }

[tool.hatch.build.targets.wheel]
packages = ["src/chain"]

[tool.ruff]
target-version = "py313"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.13"
strict = true
ignore_missing_imports = true
mypy_path = "src"
explicit_package_bases = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = ["-v", "--tb=short"]
